@@ -2,8 +2,8 @@
   "name": "@ngspot/ngx-errors",
   "version": "3.2.3",
   "peerDependencies": {
-    "@angular/common": ">=13.0.0",
-    "@angular/core": ">=13.0.0",
+    "@angular/core": ">=17.1.0",
+    "@angular/forms": ">=17.1.0",
     "rxjs": ">=7"
   },
   "dependencies": {
