@@ -6,11 +6,12 @@ export * from './lib/custom-error-state-matchers';
 export * from './lib/error-state-matchers';
 export * from './lib/error-state-matchers.service';
 export * from './lib/error.directive';
+export * from './lib/errors-base.directive';
+export * from './lib/all-errors-state.service';
 export * from './lib/errors-configuration';
 export * from './lib/errors.directive';
-export * from './lib/errors.module';
+export * from './lib/errors-declarations';
 export * from './lib/form.directive';
 export * from './lib/misc';
 export * from './lib/ngx-errors';
-export * from './lib/overridden-show-when.service';
 export * from './lib/validators';
