@@ -4,12 +4,6 @@ export class NgxError extends Error {
   }
 }
 
-export class NoParentNgxErrorsError extends NgxError {
-  constructor() {
-    super('Directive ngxError requires a parent directive ngxErrors');
-  }
-}
-
 export class ValueMustBeStringError extends NgxError {
   constructor() {
     super('Directive ngxError requires a string value');
@@ -19,7 +13,7 @@ export class ValueMustBeStringError extends NgxError {
 export class NoControlError extends NgxError {
   constructor() {
     super(
-      'Directive ngxErrors requires either control name or control instance'
+      'Directive ngxErrors requires either control name or control instance',
     );
   }
 }
@@ -39,7 +33,7 @@ export class ControlNotFoundError extends NgxError {
 export class ParentFormGroupNotFoundError extends NgxError {
   constructor(name: string) {
     super(
-      `Can't search for control "${name}" because parent FormGroup is not found`
+      `Can't search for control "${name}" because parent FormGroup is not found`,
     );
   }
 }
@@ -48,8 +42,8 @@ export class InvalidShowWhenError extends NgxError {
   constructor(showWhen: string, keys: string[]) {
     super(
       `Invalid showWhen value: ${showWhen}. Valid values are: ${keys.join(
-        ', '
-      )}`
+        ', ',
+      )}`,
     );
   }
 }
